+++ R/get_esi.R
@@ -8,7 +8,8 @@
 #'  then transformed into polygons with a small buffer area around the point.
 #' 
 #' @inheritParams get_chirps
-#' @param period an integer value for the period of ESI data, four weeks period = 1, twelve weeks = 2
+#' @param period an integer value for the period of ESI data, 
+#' four weeks period = 1, twelve weeks = 2
 #' @details
 #'  \bold{operation}: supported operations are:  
 #'  \tabular{rll}{
@@ -193,31 +194,22 @@ get_esi.sf <- function(object, dates, operation = 5, period = 1,
   
   type <- type[which(supp_type)]
   
-  n <- nrow(object)
+  nr <- dim(object)[[1]]
   
   # find the sf_column
   index <- attr(object, "sf_column")
+  
+  # get the sf column
+  lonlat <- object[[index]]
   
   if (type == "POINT") {
     
     # unlist the sf_column
     lonlat <- unlist(object[[index]])
     
-    lonlat <- matrix(lonlat,
-                     nrow = n,
-                     ncol = 2, 
-                     byrow = TRUE, 
-                     dimnames = list(1:n, c("lon","lat")))
-    
-    lonlat <- as.data.frame(lonlat)
-    
   }
   
   if (type == "POLYGON") {
-    
-    lonlat <- object[[index]]
-    
-    nl <- length(lonlat)
     
     # set centroid to validade lonlat
     lonlat <- sf::st_centroid(lonlat)
@@ -225,17 +217,17 @@ get_esi.sf <- function(object, dates, operation = 5, period = 1,
     # unlist the sf_column
     lonlat <- unlist(lonlat)
     
-    lonlat <- matrix(lonlat,
-                     nrow = nl,
-                     ncol = 2, 
-                     byrow = TRUE, 
-                     dimnames = list(1:nl, c("lon","lat")))
-    
-    lonlat <- as.data.frame(lonlat)
-    
   }
   
-  # validate lonlat to check if they are within the CHIRPS range
+  lonlat <- matrix(lonlat,
+                   nrow = nr,
+                   ncol = 2, 
+                   byrow = TRUE, 
+                   dimnames = list(seq_len(nr), c("lon","lat")))
+  
+  lonlat <- as.data.frame(lonlat)
+  
+  # validate lonlat to check if they are within the ESI range
   .validate_lonlat(lonlat, xlim = c(-180, 180), ylim = c(-50, 50))
   
   # validate and reformat dates
@@ -260,7 +252,7 @@ get_esi.sf <- function(object, dates, operation = 5, period = 1,
   if (as.sf) {
     
     result$date <- as.integer(result$date)
-    result$date <- paste0("d_",result$date)
+    result$date <- paste0("day_",result$date)
     
     result <- split(result, result$date)
     
@@ -361,7 +353,7 @@ get_esi.geojson <- function(object, dates, operation = 5, period = 1,
     
     lonlat <- as.data.frame(lonlat)
     
-    gjson <- split(object, 1:length(object))
+    gjson <- split(object, seq_along(object))
     
   }
   
@@ -388,7 +380,7 @@ get_esi.geojson <- function(object, dates, operation = 5, period = 1,
     
     result <- split(result, result$id)
     
-    object <- split(object, 1:length(object))
+    object <- split(object, seq_along(object))
     
     # add geojson properties
     result <- mapply(function(X, Y) {
