+++ R/internal_functions.R
@@ -1,10 +1,11 @@
 #' Concatenate a coordinate point data.frame into a geojson polygon
 #'
 #' Take single points from geographical coordinates and convert it into a
-#'  geojson 'Polygon' string using \code{\link[sf]{st_buffer}} 'Polygon' is the
-#'  only geojson format accepted by ClimateSERV
+#'  geojson 'Polygon' string using \code{\link[sf]{st_buffer}} 'Polygon' 
+#'  is the only geojson format accepted by ClimateSERV
 #'
-#' @param lonlat a data.frame with geographical coordinates lonlat in that order
+#' @param lonlat a data.frame with geographical coordinates lonlat in 
+#'  that order
 #' @param dist numeric; buffer distance for all \code{lonlat}
 #' @param nQuadSegs integer; number of segments per quadrant
 #' @return A list with geojson strings for each row in \code{lonlat}
@@ -15,17 +16,18 @@
 #'                      lat = runif(2, 45, 57))
 #' 
 #' gjson <- .dataframe_to_geojson(lonlat)
-#'@noRd
+#' @importFrom sf st_point st_sfc st_buffer st_write st_as_sf
+#' @noRd
 .dataframe_to_geojson <- function(lonlat,
                                   dist = 0.00001,
                                   nQuadSegs = 2L) {
-    n <- nrow(lonlat)
+    n <- dim(lonlat)[[1]]
     
     # lonlat into matrix
     lonlat <- as.matrix(lonlat)
     
     # split lonlat by rows
-    lonlat <- split(lonlat, 1:nrow(lonlat))
+    lonlat <- split(lonlat, seq_len(n))
     
     # transform into sf points
     lonlat <- lapply(lonlat, function(l) {
@@ -73,8 +75,8 @@
 #'
 #' Take single points from geographical coordinates
 #' and convert it into a geojson 'Polygon' string using
-#' \code{\link[sf]{st_buffer}} 'Polygon' is the only geojson format accepted by
-#' ClimateSERV
+#' \code{\link[sf]{st_buffer}} 'Polygon' is the only geojson format 
+#' accepted by ClimateSERV
 #'
 #' @param lonlat an object of class "sf" and geometry type "POINT"
 #' @param dist numeric; buffer distance for all \code{lonlat}
@@ -84,8 +86,8 @@
 #' # random geographic locations around bbox(10, 12, 45, 57)
 #' library("sf")
 #' set.seed(123)
-#' lonlat <- data.frame(lon = runif(2, 10, 12),
-#'                      lat = runif(2, 45, 57))
+#' lonlat <- data.frame(lon = runif(12, 10, 12),
+#'                      lat = runif(12, 45, 57))
 #' 
 #' lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))
 #' 
@@ -110,7 +112,7 @@
   type <- type[which(supp_type)]
   
   if (type == "POINT") {
-    n <- nrow(lonlat)
+    n <- dim(lonlat)[[1]]
     
     # set the buffer around the points
     lonlatb <- sf::st_buffer(lonlat,
@@ -155,22 +157,25 @@
 
 #' Send a request to ClimateSERV
 #'
-#' @param datatype integer, the unique datatype number for the dataset which
-#'  this request operates on
-#' @param begintime character, start date for processing interval, format
-#'  ("MM/DD/YYYY")
+#' @param datatype integer, the unique datatype number for the 
+#'  dataset which this request operates on
+#' @param begintime character, start date for processing interval, 
+#'  in the format ("MM/DD/YYYY")
 #' @param endtime character, end date for processing interval, format
 #'  ("MM/DD/YYYY")
 #' @param intervaltype integer, value that represents which type of time
 #'  interval to process
[…]
