+++ R/chirps.R
@@ -9,8 +9,11 @@
 #' @seealso
 #'\strong{Useful links:}
 #' \itemize{
-#' \item{Development repository: \url{https://github.com/agrobioinfoservices/chirps}}
-#' \item{Static documentation: \url{https://agrobioinfoservices.github.io/chirps/}}
-#' \item{Report bugs: \url{https://github.com/agrobioinfoservices/chirps/issues}}
+#' \item{Development repository: 
+#'  \url{https://github.com/agrobioinfoservices/chirps}}
+#' \item{Static documentation: 
+#'  \url{https://agrobioinfoservices.github.io/chirps/}}
+#' \item{Report bugs: 
+#'  \url{https://github.com/agrobioinfoservices/chirps/issues}}
 #' }
 "_PACKAGE"
