### 2020-01-09 #357 @kauedesousa
Dear @melvidoni thanks for your feedback. I've made the changes and pushed as `chirps` v0.0.5 https://github.com/agrobioinfoservices/chirps/commit/95d016fec74635290642454623f435ec6dbaa2c7
