### 2020-01-24 #357 @cvitolo
This is a very useful package, I work in the field of weather forecasting and I greatly appreciate tools that provide an interface to data services. I found your package easy to use and well documented. All your hard work made my review very easy, thanks and great job!

I have some suggestions, which I listed below, I hope you find them useful :)

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

- *README*, it seems the code in your README file is only visualised but not executed. In this case, you could keep the README.md and remove README.Rmd (as this is basically redundant).

- Your *R folder* contains a file called `sysdata.rda`. Is there a reason to keep these data there?Usually data should be placed under the root folder in a subfolder called `data`.  I would suggest to move sysdata under the `chirps/data` folder, document the datasets (documentation is currently missing for both `tapajos` and `tapajos_geom`) and load the data using `data("sysdata")` or `chirps::tapajos`. If you make this change, please change the call to `chirps:::tapajo` in `get_esi()` accordingly.

- the *man folder* contains a *figure folder*. Is this good practice? I thought the *man folder* should be used only for documentation. Would it make sense to move the figure folder under `inst`, for instance?

- *tests folder*:
  * each of your test files contain a call to library(chirps). This is superflous because you load the package in testthat.R and that should suffice. My suggestion is to load all the packages needed by the tests in testthat.R and remove the commands library(package_name) in the individual test files.
  * When you call library() sometimes you use library(package_name), other times library("package_name"). I would suggest to consistently use the latter.
  * `test-get_chirps.R`: you only test that you get the right object class but do not test the returned values, is there a reason for that? In my experience, it is very important to test the correctness of the actual data and I would suggest you to develop tests for that.
  * `test-get_esi.R`: as above, you only test that you get the right object class but do not test the returned values. I would suggest to also test the data values.
  * `test-precip_indicesl.R`:
    + the file name seems to contain an extra "l", should that not be `test-precip_indices.R`?
    + you only test the dimensions of the output data frame but not the values themselves. As, above, I would suggest to also test the data values.

- *vignettes*:
  * The file `Overview.Rmd.orig` seems redundant and can be removed.
  * It's not necessary to run twice the command `precip_indices(dat, timeseries = TRUE, intervals = 15)`, the second one can be removed.
  * When running the command`get_chirps`, I get the following warning message: `In st_buffer.sfc(st_geometry(x), dist, nQuadSegs, endCapStyle = endCapStyle,: st_buffer does not correctly buffer longitude/latitude data`. Can this warning be eliminated? Maybe adding a note in the documentation?
  * When running the command`get_chirps`, I get the following warning message: `Warning messages: 1: In st_centroid.sfc(x$geometry) : st_centroid does not give correct centroids for longitude/latitude data. 2: In st_centroid.sfc(x$geometry) : st_centroid does not give correct centroids for longitude/latitude data`. Can this warning be eliminated? Maybe adding a note in the documentation?
  * A more in-depth discussion of the functionalities included in the package will make it easier for the reader to understand if the chirps dataset is suitable for a given purpose. I would also mention that requests may take a long time to be executed. Is it feasible to use these functions to download large amount of data (for instance to perform a global scale analysis)? In general, a mention of the limitations of this package would be valuable.

- LICENSE = GPL-3
  * when you use a widely known license you should not need to add a copy of the license to your repo. The files LICENSE and LICENSE.md are redundant and can be removed.
  * When I got my packages reviewed I was made aware that GPL-3 is a strongly protective license and, if you want your package to be used widely (also commercially), MIT or Apache licenses are more suitable. I just wanted to pass on this very valuable suggestion I received.

- *inst/paper folder*:
  * it seems the code in your paper is only visualised but not executed. In this case, you could keep the paper.md and remove paper.Rmd. Also paper.pdf could be removed.
  * Fig1.svg is redundant (Fig1.png is used for rendering the paper).
  * in the paper, I would move the introduction to the CHIRPS data at the beginning as readers might not be familiar with these data.
  * in the paper you use the command `chirps:::tapajos` to load data in your sysdata.rda. This is not good practice. The `:::` operator should not be used as it exposes non-exported functionalities. If you move sysdata under the `chrips/data` folder (as suggested above), the dataset can be loaded using `data("sysdata")` or `chirps::tapajos`.
  * Towards the end of your paper you state: *Overall, these indices proved to be an excellent proxy to evaluate the climate variability using precipitation data [@DeSousa2018], the effects of climate change [@Aguilar2005], crop modelling [@Kehel2016] and to define strategies for climate adaptation [@vanEtten2019].* Maybe you could expand a bit, perhaps on the link with crop modelling?

- Lastly, I ran goodpractice::goodpractice() and got 2 messages:
  * `write unit tests for all functions, and all package code in general. 34% of code lines are covered by test cases.` This differs from what is stated on GitHub (codecv badge = ~73% code coverage). The reason might be due to the fact you skip most of your tests on cran, is this because tests take too long to run? If so, is there a way you could modify the tests so that they take less time?
  * `fix this R CMD check WARNING: Missing link or links in documentation object 'precip_indices.Rd': ‘[tidyr]{pivot_wider}’ See section 'Cross-references' in the 'Writing R Extensions' manual.` Maybe you could substitute `\code{\link[tidyr]{pivot_wider}}` with `\code{tidyr::pivot_wider()}`.
