+++ tests/testthat.R
@@ -1,4 +1,5 @@
-library(testthat)
-library(chirps)
+library("testthat")
+library("chirps")
+library("sf")
 
 test_check("chirps")

[…]
