+++ R/precip_indices.R
@@ -28,23 +28,20 @@
 #'  A. B., Mackay M., Dayanandan S.), pp. 151–174. CRC Press.
 #' 
 #' @examples
-#' 
 #' \donttest{
 #' # Three points in the Tapajos National Forest, Brazil
-#' lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
-#'                      lat = c(-2.8094, -2.8756, -3.5279))
+#' example("tapajos", package = "chirps")
 #' 
-#' dates <- c("2017-12-15","2018-05-31")
-#'  
-#' dt <- get_chirps(lonlat, dates)
+#' dt <- get_chirps(tp_point, dates = c("2017-12-15","2018-05-31"))
 #' 
-#' # take the indices for the entire period 
+#' # take the indices for the entire period
 #' precip_indices(dt, timeseries = FALSE)
 #' 
 #' # take the indices for periods of 7 days
 #' precip_indices(dt, timeseries = TRUE, intervals = 7)
 #' }
 #' @importFrom stats quantile
+#' @importFrom tibble as_tibble
 #' @export
 precip_indices <- function(object, timeseries = FALSE, intervals = NULL) {
   
@@ -136,7 +133,7 @@ precip_indices <- function(object, timeseries = FALSE, intervals = NULL) {
   
   result <- result[order(result$id), ]
   
-  result <- data.table::data.table(result)
+  result <- tibble::as_tibble(result)
   
   return(result)
 }
