+++ R/tapajos.R
@@ -1,9 +1,15 @@
 #' Tapajos National Forest
 #'
-#' Geometries for the Tapajos National Forest, a protected area in the Brazilian Amazon.
+#' Geometries for the Tapajos National Forest, a protected 
+#' area in the Brazilian Amazon
+#' \url{http://www.icmbio.gov.br/flonatapajos/}
 #' 
-#' @format A object of class sfc_POLYGON within the bbox xmin: -55.41127 ymin: -4.114584 xmax: -54.7973 ymax: -2.751706
+#' @format An object of class 'sfc_POLYGON' within the bbox 
+#' xmin: -55.41127 ymin: -4.114584 
+#' xmax: -54.7973 ymax: -2.751706
 #' @examples 
+#' library("sf")
+#' 
 #' data("tapajos")
 #' 
 #' # sample three points within the Tapajos area
@@ -14,7 +20,8 @@
 #' tp_point <- st_as_sf(tp_point)
 #' 
 #' # transform into a geojson
-#' tp_geojson <- sf_to_geojson(tp_point[1:2, ])
+#' tp_geojson <- sf_to_geojson(tp_point)
 #' 
-#' @source The data were provided by the Chico Mendes Institute \url{http://www.icmbio.gov.br/flonatapajos/}.
+#' @source The data was provided by the Chico Mendes Institute 
+#' via \url{https://www.protectedplanet.net/floresta-nacional-de-tapajos-forest}.
 "tapajos"
