+++ man/precip_indices.Rd
@@ -34,17 +34,13 @@ A dataframe with precipitation indices:
 Compute precipitation indices over a time series.
 }
 \examples{
-
 \donttest{
 # Three points in the Tapajos National Forest, Brazil
-lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
-                     lat = c(-2.8094, -2.8756, -3.5279))
+example("tapajos", package = "chirps")
 
-dates <- c("2017-12-15","2018-05-31")
- 
-dt <- get_chirps(lonlat, dates)
+dt <- get_chirps(tp_point, dates = c("2017-12-15","2018-05-31"))
 
-# take the indices for the entire period 
+# take the indices for the entire period
 precip_indices(dt, timeseries = FALSE)
 
 # take the indices for periods of 7 days
