+++ CONTRIBUTING.md
@@ -11,5 +11,5 @@ If you plan to contribute, please read our pull request process:
 1. Fork the project and clone locally.
 2. If you have large changes, please open an issue first to discuss.
 3. Documentation is generated by [roxygen2](https://CRAN.R-project.org/package=roxygen2). Please write documentation in code files and let it auto-generate documentation files. We use a recent version so documentation must be written in markdown. This [document](https://r-pkgs.org/) may help you. 
-4. Write and test your code with [good practice](https://www.r-bloggers.com/r-code-best-practices/) before submitting your change.
+4. Write and test your code with [good practice](https://github.com/mangothecat/goodpractice) before submitting your change.
 5. Push to the origin repository.

[…]
