### 2020-02-02 agrobioinfoservices/chirps#12 @Robinlovelace
Timeout when trying to get data for Peru
Minimal reproducible example below (which was originally of Honduras, hence object name) - with `withTimeout()` to show timeout (I waited over a minute but still got no response):

``` r
library(chirps)
library(dplyr)
#>
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#>
#>     filter, lag
#> The following objects are masked from 'package:base':
#>
#>     intersect, setdiff, setequal, union
library(sf)
#> Linking to GEOS 3.8.0, GDAL 3.0.2, PROJ 6.2.1
countries = spData::world
honduras = countries %>%
  filter(name_long == "Peru")
sample_sites = st_sample(honduras, 3) %>%
  st_as_sf()
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
getOption('timeout')
#> [1] 60
R.utils::withTimeout({
  dat = get_chirps(sample_sites, dates =c("2013-01-01","2013-02-01"))
}, timeout = 10)
#> Warning in st_buffer.sfc(st_geometry(x), dist, nQuadSegs, endCapStyle =
#> endCapStyle, : st_buffer does not correctly buffer longitude/latitude data
#> dist is assumed to be in decimal degrees (arc_degrees).
#> Error:
#> Something went wrong with the query, no data were returned. Please see <https://climateserv.servirglobal.net/> for potential server issues.
```

<summary>Session info</summary>

``` r
devtools::session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 3.6.2 (2019-12-12)
#>  os       Ubuntu 18.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language en_GB:en
#>  collate  en_GB.UTF-8
#>  ctype    en_GB.UTF-8
#>  tz       Europe/London
#>  date     2020-02-02
#>
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package     * version    date       lib
#>  assertthat    0.2.1      2019-03-21 [2]
#>  backports     1.1.5      2019-10-02 [1]
#>  callr         3.4.0      2019-12-09 [1]
#>  chirps      * 0.0.6      2020-02-02 [1]
#>  class         7.3-15     2019-01-01 [2]
#>  classInt      0.4-3      2019-12-04 [1]
#>  cli           2.0.1      2020-01-08 [1]
#>  crayon        1.3.4      2017-09-16 [2]
#>  crul          0.9.0      2019-11-06 [2]
#>  curl          4.3        2019-12-02 [2]
#>  DBI           1.1.0      2019-12-15 [2]
#>  desc          1.2.0      2018-05-01 [2]
#>  devtools      2.2.1      2019-09-24 [1]
#>  digest        0.6.23     2019-11-23 [1]
#>  dplyr       * 0.8.3      2019-07-04 [2]
#>  e1071         1.7-3      2019-11-26 [2]
#>  ellipsis      0.3.0      2019-09-20 [3]
#>  evaluate      0.14       2019-05-28 [2]
#>  fansi         0.4.1      2020-01-08 [1]
#>  fs            1.3.1      2019-05-06 [2]
#>  glue          1.3.1      2019-03-12 [2]
#>  highr         0.8        2019-03-20 [3]
#>  htmltools     0.4.0.9002 2020-01-25 [1]
#>  httpcode      0.2.0      2016-11-14 [2]
#>  jsonlite      1.6        2018-12-07 [2]
#>  KernSmooth    2.23-16    2019-10-15 [4]
#>  knitr         1.26       2019-11-12 [1]
#>  lwgeom        0.1-7      2019-05-06 [1]
#>  magrittr      1.5        2014-11-22 [2]
#>  memoise       1.1.0      2017-04-21 [3]
#>  pillar        1.4.3      2019-12-20 [1]
#>  pkgbuild      1.0.6      2019-10-09 [2]
#>  pkgconfig     2.0.3      2019-09-22 [2]
#>  pkgload       1.0.2      2018-10-29 [3]
#>  prettyunits   1.1.0      2020-01-09 [1]
#>  processx      3.4.1      2019-07-18 [2]
#>  ps            1.3.0      2018-12-21 [2]
#>  purrr         0.3.3      2019-10-18 [1]
#>  R.methodsS3   1.7.1      2016-02-16 [2]
#>  R.oo          1.23.0     2019-11-03 [1]
#>  R.utils       2.9.2      2019-12-08 [1]
#>  R6            2.4.1      2019-11-12 [2]
#>  Rcpp          1.0.3      2019-11-08 [1]
#>  remotes       2.1.0      2019-06-24 [1]
#>  rlang         0.4.4      2020-01-28 [1]
#>  rmarkdown     2.0.6      2020-01-13 [1]
#>  rprojroot     1.3-2      2018-01-03 [2]
#>  sessioninfo   1.1.1      2018-11-05 [3]
#>  sf          * 0.8-1      2020-01-28 [1]
#>  spData        0.3.2      2020-01-13 [1]
#>  spDataLarge   0.3.1      2020-01-12 [1]
#>  stringi       1.4.5      2020-01-11 [1]
#>  stringr       1.4.0      2019-02-10 [2]
#>  testthat      2.3.1      2019-12-01 [2]
#>  tibble        2.1.3      2019-06-06 [2]
#>  tidyselect    1.0.0      2020-01-27 [1]
#>  triebeard     0.3.0      2016-08-04 [2]
#>  units         0.6-5      2019-10-08 [1]
#>  urltools      1.7.3      2019-04-14 [2]
#>  usethis       1.5.1.9000 2020-01-31 [1]
#>  withr         2.1.2      2018-03-15 [2]
#>  xfun          0.12       2020-01-13 [1]
#>  yaml          2.2.0      2018-07-25 [3]
#>  source
#>  CRAN (R 3.6.0)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  Github (agrobioinfoservices/chirps@1858861)
#>  standard (@7.3-15)
#>  Github (r-spatial/classInt@7bc8a82)
#>  CRAN (R 3.6.2)
#>  standard (@1.3.4)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.6.2)
#>  standard (@1.2.0)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.0)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.6.0)
#>  CRAN (R 3.5.3)
#>  CRAN (R 3.5.3)
#>  Github (rstudio/htmltools@e07546c)
#>  CRAN (R 3.5.3)
#>  standard (@1.6)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.5.2)
#>  CRAN (R 3.5.0)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.5.1)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.5.3)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.5.2)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.2)
#>  Github (rstudio/rmarkdown@1c859f4)
#>  CRAN (R 3.5.3)
#>  CRAN (R 3.5.1)
#>  CRAN (R 3.6.2)
#>  Github (nowosad/spData@13d5ead)
#>  Github (nowosad/spDataLarge@1f52612)
#>  CRAN (R 3.6.2)
#>  standard (@1.4.0)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.0)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.5.3)
#>  CRAN (R 3.6.1)
#>  CRAN (R 3.6.0)
#>  Github (r-lib/usethis@dbac27b)
#>  CRAN (R 3.5.3)
#>  CRAN (R 3.6.2)
#>  CRAN (R 3.5.1)
#>
#> [1] /home/robin/R/x86_64-pc-linux-gnu-library/3.6
#> [2] /usr/local/lib/R/site-library
#> [3] /usr/lib/R/site-library
#> [4] /usr/lib/R/library
```

Great idea for a package, FYI here's some analysis of CHIRPS data that may be worth linking to, great example of using it with the `stars` package: https://github.com/loreabad6/chirps-stars
