### 2020-02-02 agrobioinfoservices/chirps#12 @kauedesousa
Hi @Robinlovelace thanks for using the package and for submitting an issue.

I am also having this problem. I think there is some problem in ClimateSERV server. I tried to run the example from their API documentation and it is not working as well.

@billyz313 , can you help us?

https://climateserv.servirglobal.net/chirps/submitDataRequest/?datatype=0&begintime=04/01/2018&endtime=04/30/2018&intervaltype=0&operationtype=5&callback=successCallback&dateType_Category=default&isZip_CurrentDataType=false&geometry={"type":"Polygon","coordinates":[[[21.533203124999996,-3.1624555302378496],[21.533203124999996,-6.489983332670647],[26.279296874999986,-5.441022303717986],[26.10351562499999,-2.635788574166625],[21.533203124999996,-3.1624555302378496]]]}

Also, thanks for sharing the documentation about the chirps-stars.
