### 2020-02-03 agrobioinfoservices/chirps#12 @kauedesousa
Hi @Robinlovelace, I just tried again using your example and it is working fine.

Let me know if it works for you as well.
