### 2020-02-03 agrobioinfoservices/chirps#12 @Robinlovelace
But if I increase the timeout to 20s it works 🎉

``` r
library(chirps)
library(dplyr)
#>
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#>
#>     filter, lag
#> The following objects are masked from 'package:base':
#>
#>     intersect, setdiff, setequal, union
library(sf)
#> Linking to GEOS 3.8.0, GDAL 3.0.2, PROJ 6.2.1
countries = spData::world
honduras = countries %>%
  filter(name_long == "Peru")
sample_sites = st_sample(honduras, 3) %>%
  st_as_sf()
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
#> although coordinates are longitude/latitude, st_intersects assumes that they are planar
getOption('timeout')
#> [1] 60
R.utils::withTimeout({
  dat = get_chirps(sample_sites, dates =c("2013-01-01","2013-02-01"))
}, timeout = 10)
#> Warning in st_buffer.sfc(st_geometry(x), dist, nQuadSegs, endCapStyle =
#> endCapStyle, : st_buffer does not correctly buffer longitude/latitude data
#> dist is assumed to be in decimal degrees (arc_degrees).
#> Getting your request...
#> Error in curl::curl_fetch_memory(x$url$url, handle = x$url$handle): Operation was aborted by an application callback
```
