+++ tests/testthat/test-get_chirps.R
@@ -1,102 +1,58 @@
 context("test-get_chirps")
 
+# load("tests/test_data.rda")
+load("../test_data.rda")
+
+# Test if get_chirps fetch the correct values,
+# for this we downloaded two points from 
+# https://climateserv.servirglobal.net/
+# and will compare it with the values retrieved by get_chirps
+
 # Test default method
-test_that("two or more points", {
+test_that("default method", {
   skip_on_cran()
-  set.seed(123)
-  lonlat <- data.frame(lon = runif(2, -55, -54),
-                       lat = runif(2, -3, -2.7))
+
+  x <- get_chirps(lonlat, dates)
   
-  dates <- c("2017-12-15", "2018-01-20")
+  x <- round(x$chirps, 2)
+
+  equal <- all(x == chirps$chirps)
   
-  
-  df <- get_chirps(lonlat, dates)
-  
-  ok <- is.data.frame(df)
-  
-  expect_true(ok)
-  
+  expect_true(equal)
+    
 })
 
+library("sf")
+# get_chirps for sf objects
+coords <- st_as_sf(lonlat, coords = c("lon","lat"))
 
-test_that("one point and other operation works", {
-  skip_on_cran()
-  set.seed(123)
-  lonlat <- data.frame(lon = runif(1, -55, -54),
-                       lat = runif(1, -3, -2.7))
-  
-  dates <- c("2017-12-15", "2018-01-20")
-  
-  
-  df <- get_chirps(lonlat, dates, operation = 2)
-  
-  ok <- is.data.frame(df)
-  
-  expect_true(ok)
-  
-})
-
-# Test S3 method for 'sf'
 test_that("sf method", {
   skip_on_cran()
-  set.seed(123)
-  lonlat <- data.frame(lon = runif(2, -55, -54),
-                       lat = runif(2, -3, -2.7))
   
-  lonlat <- st_as_sf(lonlat, coords = c("lon", "lat"))
+  y <- get_chirps(coords, dates)
   
-  dates <- c("2017-12-15", "2018-01-20")
+  y <- round(y$chirps, 2)
   
-  df <- get_chirps(lonlat, dates)
-  
-  ok <- is.data.frame(df)
-  
-  expect_true(ok)
-  
+  equal <- all(y == chirps$chirps)
+
+  expect_true(equal)
+
 })
 
-test_that("sf method return sf object", {
+# get chirps with geojson method
+geojson <- dataframe_to_geojson(lonlat)
+
+test_that("geojson method", {
   skip_on_cran()
-  set.seed(123)
-  lonlat <- data.frame(lon = runif(2, -55, -54),
-                       lat = runif(2, -3, -2.7))
   
-  lonlat <- st_as_sf(lonlat, coords = c("lon", "lat"))
-  
-  dates <- c("2017-12-15", "2018-01-20")
-  
-  df <- get_chirps(lonlat, dates, as.sf = TRUE)
-  
-  ok <- "sf" %in% class(df)
-  
-  expect_true(ok)
-  
-})
+  z <- suppressWarnings(
+    get_chirps(geojson, dates)
+    )
 
+  z <- round(z$chirps, 2)
+  
+  equal <- all(z == chirps$chirps)
 
-
-# Test errors
-test_that("points beyond lims", {
-  set.seed(123)
-  lonlat <- data.frame(lon = runif(1, -55, -54),
-                       lat = runif(1, 55, 57))
-  
-  dates <- c("2017-12-15", "2018-01-20")
-  
-  
-  expect_error(get_chirps(lonlat, dates))
-  
-})
-
-
-test_that("wrong dates", {
-  set.seed(123)
[…]
