+++ R/internal_functions.R
@@ -25,10 +25,6 @@
 #' 
 #' gjson <- dataframe_to_geojson(lonlat)
 #' 
-#' gjson <- as.character(gjson)
-#' 
-#' gsjon <- split(gjson, seq_along(gjson))
-#' 
 #' chirps:::.send_request(begintime = "12/10/2018",
 #'                        endtime = "12/26/2018",
 #'                        geometry = gjson)
@@ -54,7 +50,7 @@
       callback = "successCallback",
       dateType_Category = "default",
       isZip_CurrentDataType = "false",
-      geometry = geometry
+      geometry = toString(geometry)
     )
     
     client_request <-
