### 2020-02-05 ffbe612
add-donttest-to-pkg-examples-to-avoid-travis-issues
