+++ inst/paper/paper.bib
@@ -182,4 +182,32 @@ year = {2016}
     year = {2017},
     note = {R package version 2.3},
     url = {https://CRAN.R-project.org/package=gridExtra},
-  }
+}
+  
+  
+@article{Thomson2017,
+author = {Thomson, Madeleine C. and Ukawuba, Israel and Hershey, Christine L. and Bennett, Adam and Ceccato, Pietro and Lyon, Bradfield and Dinku, Tufa},
+doi = {10.4269/ajtmh.16-0696},
+issn = {00029637},
+journal = {American Journal of Tropical Medicine and Hygiene},
+pages = {32--45},
+publisher = {American Society of Tropical Medicine and Hygiene},
+title = {{Using rainfall and temperature data in the evaluation of national malaria control programs in Africa}},
+volume = {97},
+year = {2017}
+}
+
+@article{Horn2018,
+author = {Horn, Lindsay M. and Hajat, Anjum and Sheppard, Lianne and Quinn, Colin and Colborn, James and Zermoglio, Maria Fernanda and Gudo, Eduardo S. and Marrufo, Tatiana and Ebi, Kristie L.},
+doi = {10.3390/ijerph15040709},
+issn = {16604601},
+journal = {International Journal of Environmental Research and Public Health},
+keywords = {Climate change,Diarrheal disease,Mozambique,Precipitation,Temperature},
+month = {apr},
+number = {4},
+publisher = {MDPI AG},
+title = {{Association between precipitation and diarrheal disease in mozambique}},
+volume = {15},
+year = {2018}
+}
+

[…]
