+++ DESCRIPTION
@@ -37,7 +37,8 @@ Imports:
 Suggests:
     knitr,
     rmarkdown,
-    testthat (>= 2.1.0)
+    testthat (>= 2.1.0),
+    vcr
 Language: en-GB
 RoxygenNote: 7.0.2
 VignetteBuilder: knitr
