+++ R/tapajos.R
@@ -1,8 +1,8 @@
 #' Tapajos National Forest
 #'
 #' Geometries for the Tapajos National Forest, a protected 
-#' area in the Brazilian Amazon
-#' \url{http://www.icmbio.gov.br/flonatapajos/}
+#' area in the Brazilian 
+#' Amazon \url{http://www.icmbio.gov.br/flonatapajos/}
 #' 
 #' @format An object of class 'sfc_POLYGON' within the bbox 
 #' xmin: -55.41127 ymin: -4.114584 
@@ -25,7 +25,5 @@
 #' }
 #' 
 #' @source The data was provided by the Chico Mendes Institute via
-#' \url{
-#' https://www.protectedplanet.net/floresta-nacional-de-tapajos-forest
-#' }.
+#' \url{https://www.protectedplanet.net/}.
 "tapajos"
