### 2020-02-17 #357 @jzwart
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

I installed and reviewed _chirps v0.0.6_. See the comments above addressing the previous review that went into _v0.0.6_. The package is well-documented and easy to read. The authors have thoughtfully addressed all previous reviewer comments.

- Could you clarify what 'near-present' means in the `Overview` section of the `README.md`?
- Can you add more details for data return performance and what the user should expect for big data calls? Is the package most useful for a few point data returns or can the user get a continental data call returned in a reasonable amount of time?
- I could not run the example code successfully. I let the example code in the `README` run for about 20 min before it timed out and returned an error
```
library("chirps")

lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
                     lat = c(-2.8094, -2.8756, -3.5279))

dates <- c("2017-01-01", "2017-01-03")

dat <- get_chirps(lonlat, dates)
Something went wrong with the query, no data were returned. Please see <https://climateserv.servirglobal.net/> for potential server issues.
```
I visited the linked url and tried downloading some precip data using a custom polygon and there was no response for the ~20 minutes I waited for data to return. I'm not sure if this is a common issue with this server, but I don't feel like I can finish the review until the server is working and returns data.
I'm also not sure how I should know if there are potential server issues when I visit https://climateserv.servirglobal.net . Providing more detailed error message would help with debugging or knowing what is going on.
