### 2020-02-18 #357 @kauedesousa
Hi @jzwart I was informed by the developers of ClimateSERV that they are doing some improvement in the server. I just tried using both `chirps` and browser and it is still down for me right now. Let me know if you manage to review the package.

Otherwise what should we do @melvidoni ?
