### 2020-02-18 #357 @jzwart
Thanks @kauedesousa , `get_chirps` _was_ working for me a couple mins ago (but now down when I tried again). I'll try to finish the rest of the review now.
