### 2020-02-28 61a4f58
Enable VCR for tests requiring Internet connectivity
