+++ tests/testthat/test-get_chirps.R
@@ -4,7 +4,7 @@ context("test-get_chirps")
 load("../test_data.rda")
 
 # Test if get_chirps fetch the correct values,
-# for this we downloaded two points from 
+# for this we downloaded two points from
 # https://climateserv.servirglobal.net/
 # and will compare it with the values retrieved by get_chirps
 
@@ -12,8 +12,6 @@ load("../test_data.rda")
 
 test_that("default method", {
   vcr::use_cassette("default_method", {
-    skip_on_cran()
-    
     x <- get_chirps(lonlat, dates)
     
     x <- round(x$chirps, 2)
@@ -26,35 +24,32 @@ test_that("default method", {
 
 library("sf")
 # get_chirps for sf objects
-coords <- st_as_sf(lonlat, coords = c("lon","lat"))
+coords <- st_as_sf(lonlat, coords = c("lon", "lat"))
 
 test_that("sf method", {
-  skip_on_cran()
-  
-  y <- get_chirps(coords, dates)
-  
-  y <- round(y$chirps, 2)
-  
-  equal <- all(y == chirps$chirps)
-
-  expect_true(equal)
-
+  vcr::use_cassette("sf_method", {
+    y <- get_chirps(coords, dates)
+    
+    y <- round(y$chirps, 2)
+    
+    equal <- all(y == chirps$chirps)
+    
+    expect_true(equal)
+  })
 })
 
 # get chirps with geojson method
 geojson <- dataframe_to_geojson(lonlat)
 
+
 test_that("geojson method", {
-  skip_on_cran()
-  
-  z <- suppressWarnings(
-    get_chirps(geojson, dates)
-    )
-
-  z <- round(z$chirps, 2)
-  
-  equal <- all(z == chirps$chirps)
-
-  expect_true(equal)
+  vcr::use_cassette("geojson_method", {
+    z <- suppressWarnings(get_chirps(geojson, dates))
     
+    z <- round(z$chirps, 2)
+    
+    equal <- all(z == chirps$chirps)
+    
+    expect_true(equal)
+  })
 })
