+++ tests/fixtures/vcr_cassettes/sf_method.yml
@@ -16,7 +16,7 @@ http_interactions:
       explanation: Request fulfilled, document follows
     headers:
       status: HTTP/1.1 200 OK
-      date: Fri, 28 Feb 2020 22:15:11 GMT
+      date: Fri, 28 Feb 2020 22:26:23 GMT
       server: Apache/2.4.18 (Ubuntu)
       strict-transport-security: max-age=63072000; includeSubdomains
       x-frame-options: DENY
@@ -31,7 +31,7 @@ http_interactions:
       file: no
       string: '["Datatype not supported", "Error with begintime", "Error with endtime",
         "Error with intervaltype", "problem decoding geometry ", "Error with operationtype"]'
-  recorded_at: 2020-02-28 22:15:21 GMT
+  recorded_at: 2020-02-28 22:26:33 GMT
   recorded_with: vcr/0.4.2.94, webmockr/0.5.1.98
 - request:
     method: get
@@ -50,7 +50,7 @@ http_interactions:
       explanation: Request fulfilled, document follows
     headers:
       status: HTTP/1.1 200 OK
-      date: Fri, 28 Feb 2020 22:15:12 GMT
+      date: Fri, 28 Feb 2020 22:26:24 GMT
       server: Apache/2.4.18 (Ubuntu)
       strict-transport-security: max-age=63072000; includeSubdomains
       x-frame-options: DENY
@@ -62,8 +62,8 @@ http_interactions:
     body:
       encoding: UTF-8
       file: no
-      string: successCallback(["d0dc6842-5d86-4c3a-97f0-d7ff51de34a8"])
-  recorded_at: 2020-02-28 22:15:21 GMT
+      string: successCallback(["9e1076d0-a958-4854-a2bf-e046f0c4dc91"])
+  recorded_at: 2020-02-28 22:26:33 GMT
   recorded_with: vcr/0.4.2.94, webmockr/0.5.1.98
 - request:
     method: get
@@ -82,7 +82,7 @@ http_interactions:
       explanation: Request fulfilled, document follows
     headers:
       status: HTTP/1.1 200 OK
-      date: Fri, 28 Feb 2020 22:15:13 GMT
+      date: Fri, 28 Feb 2020 22:26:25 GMT
       server: Apache/2.4.18 (Ubuntu)
       strict-transport-security: max-age=63072000; includeSubdomains
       x-frame-options: DENY
@@ -97,7 +97,7 @@ http_interactions:
       file: no
       string: '["Datatype not supported", "Error with begintime", "Error with endtime",
         "Error with intervaltype", "problem decoding geometry ", "Error with operationtype"]'
-  recorded_at: 2020-02-28 22:15:21 GMT
+  recorded_at: 2020-02-28 22:26:33 GMT
   recorded_with: vcr/0.4.2.94, webmockr/0.5.1.98
 - request:
     method: get
@@ -116,7 +116,7 @@ http_interactions:
       explanation: Request fulfilled, document follows
     headers:
       status: HTTP/1.1 200 OK
-      date: Fri, 28 Feb 2020 22:15:14 GMT
+      date: Fri, 28 Feb 2020 22:26:25 GMT
       server: Apache/2.4.18 (Ubuntu)
       strict-transport-security: max-age=63072000; includeSubdomains
       x-frame-options: DENY
@@ -128,8 +128,8 @@ http_interactions:
     body:
       encoding: UTF-8
       file: no
-      string: successCallback(["aa15ca8d-8ec7-4498-b9bf-f9febcfd9e94"])
-  recorded_at: 2020-02-28 22:15:21 GMT
[…]

[…]
