+++ .travis.yml
@@ -5,7 +5,7 @@ cache: packages
 r:
   - release
   - devel
-  #- oldrel
+  - oldrel
 
 warnings_are_errors: true
 
