### 2020-05-05 1aad5db
fixes a few code formatting things
