+++ vignettes/precompile.R
@@ -19,4 +19,4 @@ resources <-
   list.files("vignettes/", pattern = ".png$", full.names = TRUE)
 file.copy(from = resources,
           to = "docs",
-          overwrite =  TRUE)
+          overwrite = TRUE)
