+++ R/get_chirps.R
@@ -50,21 +50,22 @@
 #' \code{\link[sf]{sf}}, please look sf documentation for 
 #' possible issues.
 #' @examples
-#' \donttest{ 
-#' 
+#' \donttest{
 #' lonlat <- data.frame(lon = c(-55.0281,-54.9857),
 #'                      lat = c(-2.8094, -2.8756))
 #' 
 #' dates <- c("2017-12-15", "2017-12-31")
 #' 
 #' dt <- get_chirps(lonlat, dates)
-#'  
+#' 
+#' dt
+#' 
 #' ############################################
 #' 
 #' # S3 method for objects of class 'sf'
 #' 
 #' ## geometry 'POINT'
-#' example("tapajos", package = "chirps")
+#' data("tapajos", package = "chirps")
 #' 
 #' dates <- c("2017-12-15", "2017-12-31")
 #' 
@@ -98,8 +99,8 @@
 #' pol <- st_as_sf(pol)
 #' 
 #' dt <- get_chirps(pol, dates = c("2018-01-01", "2018-01-20"))
+#' }
 #' 
-#' }  
 #' @importFrom sf st_centroid read_sf st_geometry_type
 #' @importFrom methods addNextMethod asMethodDefinition assignClassDef
 #' @export
