+++ inst/paper/paper.bib
@@ -16,11 +16,11 @@ year = {2015}
 
 
 @Manual{RCoreTeam,
-    title = {R: A Language and Environment for Statistical Computing},
+    title = {R: A Language and Environment for Statistical Computing. version 4.0.0.},
     author = {{R Core Team}},
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
-    year = {2019},
+    year = {2020},
     url = {https://www.R-project.org/},
 }
 
