+++ R/get_esi.R
@@ -38,10 +38,6 @@
 #' 
 #' @references
 #' ClimateSERV \url{https://climateserv.servirglobal.net}
-#' 
-#' Evaporative Stress Index
-#' \url{http://catalogue.servirglobal.net/Product?product_id=198}
-#' 
 #' @examples
 #' \donttest{
 #' 
