+++ R/chirps.R
@@ -1,10 +1,16 @@
 #' API Client for CHIRPS
 #'
 #' @name chirps
-#' @author Kauê de Sousa and Adam H. Sparks and William Ashmall and Jacob van
-#'  Etten and Svein Ø. Solberg
+#' @author Kauê de Sousa and Adam H. Sparks
 #' @docType package
-#'
+#' @import methods
+#' @note 
+#' While chirps does not redistribute the data or provide it in any way, 
+#'  we encourage users to cite Funk et al. (2015) when using the CHIRPS data.
+#' 
+#' Funk C., et al. (2015). Scientific Data, 2, 150066. 
+#' \url{https://doi.org/10.1038/sdata.2015.66}
+#' 
 #' @seealso
 #'\strong{Useful links:}
 #' \itemize{
@@ -14,5 +20,9 @@
 #'  \url{https://agrobioinfoservices.github.io/chirps/}}
 #' \item{Report bugs: 
 #'  \url{https://github.com/agrobioinfoservices/chirps/issues}}
+#' \item{CHIRPS website: 
+#'  \url{https://www.chc.ucsb.edu/data/chirps}}
+#' \item{ClimateSERV website: 
+#'  \url{https://climateserv.servirglobal.net/}}
 #' }
 "_PACKAGE"
