+++ man/chirps.Rd
@@ -17,6 +17,13 @@ API Client for the Climate Hazards Group InfraRed Precipitation
   Requests from large time series (> 10 years) and large geographic coverage 
   (global scale) may take several minutes.
 }
+\note{
+While chirps does not redistribute the data or provide it in any way, 
+ we encourage users to cite Funk et al. (2015) when using the CHIRPS data.
+
+Funk C., et al. (2015). Scientific Data, 2, 150066. 
+\url{https://doi.org/10.1038/sdata.2015.66}
+}
 \seealso{
 \strong{Useful links:}
 \itemize{
@@ -26,9 +33,12 @@ API Client for the Climate Hazards Group InfraRed Precipitation
  \url{https://agrobioinfoservices.github.io/chirps/}}
 \item{Report bugs: 
  \url{https://github.com/agrobioinfoservices/chirps/issues}}
+\item{CHIRPS website: 
+ \url{https://www.chc.ucsb.edu/data/chirps}}
+\item{ClimateSERV website: 
+ \url{https://climateserv.servirglobal.net/}}
 }
 }
 \author{
-Kauê de Sousa and Adam H. Sparks and William Ashmall and Jacob van
- Etten and Svein Ø. Solberg
+Kauê de Sousa and Adam H. Sparks
 }
