+++ man/get_chirps.Rd
@@ -83,7 +83,7 @@ dt
 ############################################
 
 # S3 method for objects of class 'sf'
-
+example("tapajos", "chirps")
 ## geometry 'POINT'
 data("tapajos", package = "chirps")
 
