+++ NAMESPACE
@@ -12,6 +12,7 @@ export(get_chirps)
 export(get_esi)
 export(precip_indices)
 export(sf_to_geojson)
+import(methods)
 importFrom(crul,HttpClient)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,toJSON)
