+++ .travis.yml
@@ -11,13 +11,11 @@ warnings_are_errors: false
 before_install:
   - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
   - sudo apt-get --yes --force-yes update -qq
-  # sf dependencies:
   - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
-apt_packages:
-  - libgdal-dev
-  - libproj-dev
+  - sudo ldconfig
 r_packages:
   - lwgeom
+  - rgdal
 after_success:
   - Rscript -e 'covr::codecov()'
 
