+++ .travis.yml
@@ -8,14 +8,6 @@ r:
 os:
   - linux
 warnings_are_errors: false
-before_install:
-  - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
-  - sudo apt-get --yes --force-yes update -qq
-  - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
-  - sudo ldconfig
-r_packages:
-  - lwgeom
-  - rgdal
 after_success:
   - Rscript -e 'covr::codecov()'
 
