+++ .travis.yml
@@ -1,13 +1,29 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 language: r
 cache: packages
-r:
-  - release
-  - devel
-  - oldrel
-os:
-  - linux
-warnings_are_errors: false
+latex: false
+sudo: required
+
+matrix:
+  include:
+    - os: linux
+      r: release
+      before_install:
+        - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+        - sudo apt-get --yes --force-yes update -qq
+        # units/udunits2 dependency:
+        - sudo apt-get install --yes libudunits2-dev
+        # sf dependencies:
+        - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev
+    - os: linux
+      r: oldrel
+      before_install:
+        - sudo add-apt-repository ppa:ubuntugis/ubuntugis-unstable --yes
+        - sudo apt-get --yes --force-yes update -qq
+        # units/udunits2 dependency:
+        - sudo apt-get install --yes libudunits2-dev
+        # sf dependencies:
+        - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev
+    - os: osx
+
 after_success:
   - Rscript -e 'covr::codecov()'
-
