+++ inst/paper/paper.Rmd
@@ -1,88 +1,0 @@
----
-title: 'chirps: API Client for the CHIRPS Precipitation Data in R'
-tags:
-- CHIRPS
-- climate data
-- climatology
-- earth science
-- evapotranspiration
-- precipitation data
-- R
-- reproducibility
-- weather data
-authors:
-  - name: Kauê de Sousa
-    orcid: 0000-0002-7571-7845
-    affiliation: "1, 2"
-  - name: Adam H. Sparks
-    orcid: 0000-0002-0061-8359
-    affiliation: 3
-  - name: William Ashmall
-    affiliation: 4
-  - name: Jacob van Etten
-    orcid: 0000-0001-7554-2558
-    affiliation: 2
-  - name: Svein Ø. Solberg
-    orcid: 0000-0002-4491-4483
-    affiliation: 1
-affiliations:
-  - name: Department of Agricultural Sciences, Inland Norway University of Applied Sciences, Hamar, Norway
-    index: 1
-  - name: Bioversity International, Rome, Italy
-    index: 2
-  - name: Centre for Crop Health, University of Southern Queensland, Toowoomba, Australia
-    index: 3
-  - name: Universities Space Research Association, National Aeronautics and Space Administration (NASA), Huntsville, USA
-    index: 4
-citation_author: de Sousa et. al.
-date: "`r format(Sys.time(), '%d %B %Y')`"
-year: 2020
-bibliography: paper.bib
-output: rticles::joss_article
-journal: JOSS
----
-
-# Summary
-
-The *chirps* package provides functionalities for reproducible analysis in R [@RCoreTeam] using the CHIRPS [@Funk2015] and ESI [@esi] data. CHIRPS is daily precipitation data set developed by the Climate Hazards Group [@Funk2015] for high resolution precipitation gridded data. Spanning 50$^{\circ}$ S to 50$^{\circ}$ N (and all longitudes) and ranging from 1981 to near-present (normally with a 45 day lag), CHIRPS incorporates 0.05 arc-degree resolution satellite imagery, and in-situ station data to create gridded precipitation time series for trend analysis and seasonal drought monitoring [@Funk2015]. The evaporative stress index (ESI) data describes temporal anomalies in evapotranspiration produced weekly at 0.25 arc-degree resolution for the entire globe [@Anderson2011]. The ESI data is based on satellite observations of land surface temperature, which are used to estimate water loss due to evapotranspiration (the sum of evaporation and plant transpiration from the Earth's land and ocean surface to the atmosphere). The ESI data is available from 2001 to near-present. When using these data sets in publications please cite @Funk2015 for CHIRPS and @esi for ESI.
-
-# Implementation
-
[…]
