### 2020-05-23 ea8bac7
Update a few things in documentation. Nothing major.
