+++ R/tapajos.R
@@ -1,8 +1,7 @@
 #' Tapajos National Forest
 #'
 #' Geometries for the Tapajos National Forest, a protected 
-#' area in the Brazilian 
-#' Amazon \url{http://www.icmbio.gov.br/flonatapajos/}
+#' area in the Brazilian Amazon \url{http://www.icmbio.gov.br/flonatapajos/}
 #' 
 #' @format An object of class 'sfc_POLYGON' within the bounding box 
 #' xmin: -55.41127 ymin: -4.114584 
