### 2020-05-24 b85b6af
Add GitHub workflow to check package
