+++ .github/workflows/check_standard.yaml
@@ -2,10 +2,11 @@ on:
   push:
     branches:
       - master
+      - devel
   pull_request:
     branches:
       - master
-
+      - devel
 name: R-CMD-check
 
 jobs:
