### 2020-05-24 ef7f1ea
Revert to use macOS for R devel (see if it works
