### 2020-05-24 a0adefd
Use sf yaml file to install system files for macOS
