+++ .github/workflows/check_standard.yaml
@@ -37,7 +37,7 @@ jobs:
 
       - uses: r-lib/actions/setup-pandoc@master
       
-            - name: Brew and macOS config
+      - name: Brew and macOS config
         if: runner.os == 'macOS'
         run: |
           brew install pkg-config
