### 2020-05-24 9db00e9
Rename 4.0 to 'release'
