### 2020-06-12 48cf00a
Correct spelling of "extreme" for extreme_wet_days()
