### 2020-06-25 f7ee261
Fix error message with documentation links
