+++ R/st_to_geojson.R
@@ -2,7 +2,7 @@
 #'
 #' Take single points from geographical coordinates
 #' and convert it into a geojson 'Polygon' string using
-#' \code{\link[sf]{st_buffer}}.
+#'  \code{\link[sf:st_buffer]{st_buffer}}.
 #'
 #' @param lonlat an object of class 'sf' and geometry type 'POINT' or 'POLYGON'
 #' @param dist numeric, buffer distance for all \code{lonlat}
