### 2020-06-26 910efbe
try again to fix R-devel docu error
