+++ man/dataframe_to_geojson.Rd
@@ -4,7 +4,7 @@
 \alias{dataframe_to_geojson}
 \title{Concatenate a data.frame into a geojson polygon}
 \usage{
-dataframe_to_geojson(lonlat, dist = 1e-05, nQuadSegs = 2L, ...)
+dataframe_to_geojson(lonlat, dist = 0.00001, nQuadSegs = 2L, ...)
 }
 \arguments{
 \item{lonlat}{a data.frame with geographical coordinates lonlat in 
@@ -21,7 +21,7 @@ An object of class 'geojson' for each row in \code{lonlat}
 }
 \description{
 Take single points from geographical coordinates and convert it into a
-geojson 'Polygon' string using \code{\link[sf:st_buffer]{st_buffer}}.
+geojson 'Polygon' string using \code{\link[sf:geos_unary]{st_buffer}}.
 }
 \examples{
 \donttest{
