### 2020-06-26 4a27ba9
Remove Travis badge
