### 2020-06-29 #357 @cvitolo
Hi @melvidoni, I’m very happy with the revised version. All my suggestions have been addressed.

@kauedesousa great job!
