### 2020-06-29 #357 @jzwart
Nice job on this package, @melvidoni, and thank you for the careful consideration of all the reviews. I find the revised package very easy to use and the descriptions in the vignettes & README useful. Thank you for adding in the time out for `.GET()` - I know that's probably a rare case with server issues, but I think adds to the functionality of the package. All my suggestions have been addressed.
