### 2020-06-29 #357 @melvidoni
Approved! Thanks @kauedesousa  for submitting and @cvitolo @jzwart  for your reviews!

- [ ] Submit to JOSS at <https://joss.theoj.org/papers/new>, using the rOpenSci GitHub repo URL. When a JOSS "PRE REVIEW" issue is generated for your paper, add the comment: `This package has been reviewed by rOpenSci: https://LINK.TO/THE/REVIEW/ISSUE, @ropensci/editors`
