### 2020-06-29 #357 @kauedesousa
I already need. Can you assign the repo to the team chirps that you set up. I skipped this step and now I can't go into the repo settings.
