### 2020-06-30 #357 @jzwart
@kauedesousa, no need to acknowledge me as a reviewer in the package description, but thank you for asking!
