+++ R/chirps.R
@@ -15,11 +15,11 @@
 #'\strong{Useful links:}
 #' \itemize{
 #' \item{Development repository: 
-#'  \url{https://github.com/agrobioinfoservices/chirps}}
+#'  \url{https://github.com/ropensci/chirps}}
 #' \item{Static documentation: 
-#'  \url{https://agrobioinfoservices.github.io/chirps/}}
+#'  \url{https://docs.ropensci.org/chirps/}}
 #' \item{Report bugs: 
-#'  \url{https://github.com/agrobioinfoservices/chirps/issues}}
+#'  \url{https://github.com/ropensci/chirps/issues}}
 #' \item{CHIRPS website: 
 #'  \url{https://www.chc.ucsb.edu/data/chirps}}
 #' \item{ClimateSERV website: 
