+++ NEWS.md
@@ -1,3 +1,9 @@
+chirps 0.1.0 (2020-06-01)
+=========================
+
+* rOpenSci release version
+
+
 chirps 0.0.8 (2020-05-22)
 =========================
 
