### 2020-07-01 9e92a6c
Update test_coverage.yaml
