+++ .github/workflows/test_coverage.yaml
@@ -12,7 +12,7 @@ jobs:
   test-coverage:
     runs-on: macOS-latest
     env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      GITHUB_PAT: ${{ secrets.chirps_TOKEN }}
     steps:
       - uses: actions/checkout@v2
 
