+++ README.md
@@ -1,9 +1,10 @@
 <!-- badges: start -->
 [![peer-review](https://badges.ropensci.org/357_status.svg)](https://github.com/ropensci/software-review/issues/357)
+[![R build status](https://github.com/ropensci/chirps/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/chirps/actions)
 [![CRAN status](https://www.r-pkg.org/badges/version/chirps)](https://cran.r-project.org/package=chirps)
-![R-CMD-check](https://github.com/ropensci/chirps/workflows/R-CMD-check/badge.svg)
 [![codecov](https://codecov.io/gh/ropensci/chirps/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/chirps)
 [![Project Status](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![DOI](https://zenodo.org/badge/225693680.svg)](https://zenodo.org/badge/latestdoi/225693680)
 <!-- badges: end -->
 
 # *chirps*: API Client for CHIRPS <img align="right" src="man/figures/logo.png">
