+++ .github/workflows/test_coverage.yaml
@@ -37,7 +37,7 @@ jobs:
         run: |
           install.packages(c("remotes"))
           remotes::install_deps(dependencies = TRUE)
-          remotes::install_cran("covr")
+          remotes::install_cran(c("covr", "tibble"))
         shell: Rscript {0}
 
       - name: Test coverage
