### 2020-07-01 5cc8ac4
update-contributing-and-delete-code-of-conduct
