+++ CONTRIBUTING.md
@@ -1,15 +1,54 @@
-# Contributing
+# CONTRIBUTING #
 
-When contributing to **chirps**, please first discuss the change you wish to make via [issue](https://github.com/ropensci/chirps/issues) before making a change and pull request. Please submit questions, bug reports, and requests in the issues tracker.
+### Fixing typos
 
-Note that we have a [code of conduct](https://ropensci.github.io/chirps/CODE_OF_CONDUCT.html), please follow it in all your interactions with the project's team.
+Small typos or grammatical errors in documentation may be edited directly using
+the GitHub web interface, so long as the changes are made in the _source_ file.
 
-If you plan to contribute, please read our pull request process:
+*  YES: you edit a roxygen comment in a `.R` file below `R/`.
+*  NO: you edit an `.Rd` file below `man/`.
 
-### Pull Request Process
+### Prerequisites
 
-1. Fork the project and clone locally.
-2. If you have large changes, please open an issue first to discuss.
-3. Documentation is generated by [roxygen2](https://CRAN.R-project.org/package=roxygen2). Please write documentation in code files and let it auto-generate documentation files. We use a recent version so documentation must be written in markdown. This [document](https://r-pkgs.org/) may help you. 
-4. Write and test your code with [good practice](https://github.com/mangothecat/goodpractice) before submitting your change.
-5. Push to the origin repository.
+Before you make a substantial pull request, you should always file an issue and
+make sure someone from the team agrees that it’s a problem. If you’ve found a
+bug, create an associated issue and illustrate the bug with a minimal 
+[reprex](https://www.tidyverse.org/help/#reprex).
+
+### Pull request process
+
+*  We recommend that you create a Git branch for each pull request (PR).  
+*  Look at the build status before and after making changes.
+The `README` should contain badges for any continuous integration services used
+by the package.  
+*  We recommend the tidyverse [style guide](http://style.tidyverse.org).
+You can use the [styler](https://CRAN.R-project.org/package=styler) package to
+apply these styles, but please don't restyle code that has nothing to do with 
+your PR.  
+*  We use [roxygen2](https://cran.r-project.org/package=roxygen2).  
+*  We use [testthat](https://cran.r-project.org/package=testthat). Contributions
+with test cases included are easier to accept.  
+*  For user-facing changes, add a bullet to the top of `NEWS.md` below the
+current development version header describing the changes made followed by your
+GitHub username, and links to relevant issue(s)/PR(s).
+
+See rOpenSci [contributing guide](https://devguide.ropensci.org/contributingguide.html) for further details.
+
+### Code of Conduct
+
+Please note that the *chirps* project is released with a a [Contributor Code of Conduct](https://ropensci.org/code-of-conduct/). By contributing to this project, you agree to abide by its terms.
+
+### Discussion forum
+
+Check out our [discussion forum](https://discuss.ropensci.org) if you think your issue requires a longer form discussion.
+
+### Prefer to Email? 
+
+Email the person listed as maintainer in the `DESCRIPTION` file of this repo.
+
+Though note that private discussions over email don't help others - of course email is totally warranted if it's a sensitive problem of any kind.
+
+### Thanks for contributing!
+
+This contributing guide is adapted from the tidyverse contributing guide available at https://raw.githubusercontent.com/r-lib/usethis/master/inst/templates/tidy-contributing.md 
+
