### 2020-07-01 #357 @kauedesousa
Hi @melvidoni , the to-do list is completed. Let me know if there is something else to do. My next step is submit this to CRAN.
