### 2020-07-01 96daec4
Fix errors when running R CMD check locally
