+++ vignettes/Overview.Rmd
@@ -14,9 +14,10 @@ author:
   affiliation: Department of Agricultural Sciences, Inland Norway University, Hamar, Norway
 output: html_document
 vignette: >
-  %\VignetteIndexEntry{Overview}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteEngine{knitr::rmarkdown_notangle}
+vignette: >
+  %\VignetteEngine{knitr::knitr}
+  %\VignetteIndexEntry{Introduction to chirps}
+  %\usepackage[UTF-8]{inputenc}
 bibliography: ["chirps.bib"]
 csl: citation_style.csl
 ---
@@ -37,172 +38,21 @@ The *Tapajós* National Forest is a protected area in the Brazilian Amazon. Loca
 
 
 
-```r
-library("chirps")
-library("sf")
-
-example("tapajos", package = "chirps")
-
-dt <- get_chirps(tp_point, dates = c("2013-01-01","2018-12-31"))
-```
-
-
-The function `get_chirps()` returns a data.frame which inherits the classes 'chirps' and 'chirps_df', where each id represents the index for the rows in the in-putted 'object'. 
-
-
-```r
-dt
-#>          id    lon   lat       date chirps
-#>       <int>  <dbl> <dbl>     <date>  <dbl>
-#> 1:        1 -55.03 -3.80 2013-01-01   0.00
-#> 2:        1 -55.03 -3.80 2013-01-02  12.05
-#> 3:        1 -55.03 -3.80 2013-01-03  24.11
-#> 4:        1 -55.03 -3.80 2013-01-04   0.00
-#> 5:        1 -55.03 -3.80 2013-01-05   0.00
-#> ---                                       
-#> 6569:     3 -55.03 -3.41 2018-12-27   0.00
-#> 6570:     3 -55.03 -3.41 2018-12-28   0.00
-#> 6571:     3 -55.03 -3.41 2018-12-29   0.00
-#> 6572:     3 -55.03 -3.41 2018-12-30   0.00
-#> 6573:     3 -55.03 -3.41 2018-12-31   0.00
-```
-
-With `precip_indices()` is possible to assess how the precipitation changes across a time series using precipitation variability indices[@Aguilar2005]. Here, we take the indices for intervals of 15 days and compute the indices for the time series (from Jan-2013 to Dec-2018).
-
-
-```r
-p_ind <- precip_indices(dt, timeseries = TRUE, intervals = 15)
-
-p_ind
-#>          id       date    lon   lat  index  value
-#>       <int>     <date>  <dbl> <dbl>  <chr>  <dbl>
-#> 1:        1 2013-01-01 -55.03 -3.80   MLDS   7.00
-#> 2:        1 2013-01-01 -55.03 -3.80   MLWS   2.00
-#> 3:        1 2013-01-01 -55.03 -3.80  R10mm   1.00
-#> 4:        1 2013-01-01 -55.03 -3.80  R20mm   3.00
-#> 5:        1 2013-01-01 -55.03 -3.80 Rx1day  47.37
-#> ---                                              
-#> 3446:     3 2018-12-16 -55.03 -3.41 Rx5day  53.90
-#> 3447:     3 2018-12-16 -55.03 -3.41   R95p  34.53
-#> 3448:     3 2018-12-16 -55.03 -3.41   R99p  34.53
-#> 3449:     3 2018-12-16 -55.03 -3.41 Rtotal  80.49
-#> 3450:     3 2018-12-16 -55.03 -3.41   SDII  13.42
-```
-
-The function `precip_indices()` returns a data.frame with the precipitation indices. Each date corresponds to the first day in the time series intervals as defined by the argument 'intervals'. When `timeseries = FALSE` the function returns a single precipitation index for the entire time series.
-
[…]
