+++ R/get_imerg.R
@@ -1,0 +1,328 @@
+#' Get Integrated Multisatellite Retrievals for GPM (IMERG) data
+#' 
+#'  The IMERG dataset provides near-real time global observations of 
+#'  rainfall at 10km resolution, which can be used to estimate total
+#'  rainfall accumulation from storm systems and quantify the intensity 
+#'  of rainfall and flood impacts from tropical cyclones and other storm 
+#'  systems. IMERG is a daily precipitation dataset available from 2015 
+#'  to present within the latitudes 70 and -70.
+#' 
+#' @inheritParams get_chirps
+#' @details
+#'  \bold{operation}: supported operations are:  
+#'  \tabular{rll}{
+#'  \bold{operation}      \tab    \tab \bold{value}\cr
+#'  max                   \tab =  \tab 0\cr
+#'  min                   \tab =  \tab 1\cr
+#'  median                \tab =  \tab 2\cr
+#'  sum                   \tab =  \tab 4\cr
+#'  average               \tab =  \tab 5 (\emph{default value})\cr
+#'  }
+#' 
+#' \bold{dist}: numeric, buffer distance for each \code{object} coordinate
+#' 
+#' \bold{nQuadSegs}: integer, number of segments per buffer quadrant
+#' 
+#' @return A data frame of \acronym{imerg} data:
+#' \item{id}{the index for the rows in \code{object}}
+#' \item{dates}{the dates from which imerg was requested}
+#' \item{lon}{the longitude as provided in \code{object}}
+#' \item{lat}{the latitude as provided in \code{object}}
+#' \item{imerg}{the IMERG value}
+#' 
+#' @references
+#' ClimateSERV \url{https://climateserv.servirglobal.net}
+#' 
+#' NASA IMERG  \url{https://disasters.nasa.gov/instruments/imerg}
+#' @examples
+#' \donttest{
+#' lonlat <- data.frame(lon = c(-55.0281,-54.9857),
+#'                      lat = c(-2.8094, -2.8756))
+#' 
+#' dates <- c("2017-12-15", "2017-12-31")
+#' 
+#' dt <- get_imerg(lonlat, dates)
+#' 
+#' dt
+#' } 
+#' @importFrom sf st_centroid read_sf st_geometry_type
+#' @export
+get_imerg <- function(object, dates, operation = 5, ...) {
+  
+  UseMethod("get_imerg")
+  
+}
+
+#' @rdname get_imerg
+#' @export
+get_imerg.default <- function(object, dates, operation = 5, ...) {
+  
+  object <- as.data.frame(object)
+  
+  # validate lonlat to check if they are within the IMERG range lat -60, 60
+  .validate_lonlat(object, xlim = c(-180, 180), ylim = c(-70, 70))
+  
+  # validate dates
+  dates_inter <- .reformat_dates(dates, availability = c("2015-01-01", "0"))
+  
+  # get geojson strings from data.frame
+  gj <- dataframe_to_geojson(object, ...)
+  
+  class(gj) <- "character"
+  
+  gj <- split(gj, seq_along(gj))
+  
+  result <- .GET(gjson = gj, 
+                 dates = dates_inter, 
+                 operation = operation, 
+                 datatype = 26)
+  
+  names(result)[names(result) == "value"] <- "imerg"
+  
+  object$id <- rownames(object)
+  
+  result <- merge(result, object, by = "id", all.y = TRUE)
+  
+  names(result)[3:5] <- c("imerg", "lon", "lat")
+  
+  result <- result[, c("id", "lon", "lat", "date", "imerg")]
+  
+  result <- as.data.frame(result, stringsAsFactors = FALSE)
+  
+  class(result) <- union(c("chirps_df", "chirps"), class(result))
+  
[…]
