+++ NAMESPACE
@@ -6,10 +6,14 @@ S3method(get_chirps,sf)
 S3method(get_esi,default)
 S3method(get_esi,geojson)
 S3method(get_esi,sf)
+S3method(get_imerg,default)
+S3method(get_imerg,geojson)
+S3method(get_imerg,sf)
 S3method(print,chirps_df)
 export(dataframe_to_geojson)
 export(get_chirps)
 export(get_esi)
+export(get_imerg)
 export(precip_indices)
 export(sf_to_geojson)
 import(methods)
