+++ inst/paper/paper.bib
@@ -211,3 +211,13 @@ volume = {15},
 year = {2018}
 }
 
+@misc{Huffman2014,
+author = {Huffman, G. and Bolvin, D. and Braithwaite, D. and Hsu, K. and Joyce, R. and Xie, P.},
+publisher = {NASA's Precipitation Processing Center},
+title = {{Integrated Multi-satellitE Retrievals for GPM (IMERG), version 4.4.}},
+url = {ftp://arthurhou.pps.eosdis.nasa.gov/gpmdata/},
+year = {2014}
+}
+
+
+
