+++ inst/paper/paper.bib
@@ -169,6 +169,7 @@ chapter = {13},
 editor = {Bari, Abdallah and Damania, Ardeshir B. and Mackay, Michael and Dayanandan, Selvadurai},
 isbn = {9781498730136},
 mendeley-groups = {tricotgenomic,gosset},
+doi = {10.1201/b19518-17},
 pages = {151--174},
 publisher = {CRC Press},
 title = {{Identifying Climate Patterns during the Crop-Growing Cycle from 30 Years of CIMMYT Elite Spring Wheat International Yield Trials}},
