+++ inst/paper/paper.bib
@@ -110,7 +110,7 @@ year = {2018}
   
 
 @Manual{crul,
-    title = {crul: HTTP Client},
+    title = {{crul}: HTTP Client},
     author = {Scott Chamberlain},
     year = {2019},
     note = {R package version 0.7.4},
@@ -126,7 +126,7 @@ journal = {Journal of Climate},
 month = {apr},
 number = {8},
 pages = {2025--2044},
-title = {{Evaluation of Drought Indices Based on Thermal Remote Sensing of Evapotranspiration over the Continental United States}},
+title = {{{E}valuation of {D}rought {I}ndices {B}ased on {T}hermal {R}emote {S}ensing of {E}vapotranspiration over the {C}ontinental {U}nited {S}tates}},
 url = {http://journals.ametsoc.org/doi/10.1175/2010JCLI3812.1},
 volume = {24},
 year = {2011}
@@ -156,7 +156,7 @@ doi = {10.21105/joss.01686},
 journal = {Journal of Open Source Software},
 number = {43},
 pages = {1686},
-title = {{Welcome to the Tidyverse}},
+title = {{{W}elcome to the {T}idyverse}},
 volume = {4},
 year = {2019}
 }
@@ -164,7 +164,7 @@ year = {2019}
 
 @incollection{Kehel2016,
 author = {Kehel, Z. and Crossa, J. and Reynolds, M.},
-booktitle = {Applied Mathematics and Omics to Assess Crop Genetic Resources for Climate Change Adaptive Traits},
+booktitle = {{A}pplied {M}athematics and {O}mics to {A}ssess {C}rop {G}enetic {R}esources for {C}limate {C}hange {A}daptive {T}raits},
 chapter = {13},
 editor = {Bari, Abdallah and Damania, Ardeshir B. and Mackay, Michael and Dayanandan, Selvadurai},
 isbn = {9781498730136},
@@ -172,13 +172,13 @@ mendeley-groups = {tricotgenomic,gosset},
 doi = {10.1201/b19518-17},
 pages = {151--174},
 publisher = {CRC Press},
-title = {{Identifying Climate Patterns during the Crop-Growing Cycle from 30 Years of CIMMYT Elite Spring Wheat International Yield Trials}},
+title = {{{I}dentifying {C}limate {P}atterns during the {C}rop-{G}rowing {C}ycle from 30 {Y}ears of {CIMMYT} {E}lite {S}pring {W}heat {I}nternational {Y}ield {T}rials}},
 year = {2016}
 }
 
 
 @Manual{gridExtra,
-    title = {gridExtra: Miscellaneous Functions for "Grid" Graphics},
+    title = {{gridExtra}: {M}iscellaneous {F}unctions for "{G}rid" {G}raphics},
     author = {Baptiste Auguie},
     year = {2017},
     note = {R package version 2.3},
