### 2020-07-01 f93650d
update-bib-and-minor-edits-in-the-paper-file
