+++ inst/paper/paper.bib
@@ -2,7 +2,6 @@
 author = {Funk, Chris and Peterson, Pete and Landsfeld, Martin and Pedreros, Diego and Verdin, James and Shukla, Shraddhanand and Husak, Gregory and Rowland, James and Harrison, Laura and Hoell, Andrew and Michaelsen, Joel},
 doi = {10.1038/sdata.2015.66},
 journal = {Scientific Data},
-keywords = {chirps,rain},
 mendeley-groups = {tricotgenomic},
 mendeley-tags = {chirps,rain},
 month = {dec},
@@ -50,8 +49,6 @@ author = {Aguilar, E. and Peterson, T. C. and Obando, P. Ram{\'{i}}rez and Fruto
 doi = {10.1029/2005JD006119},
 issn = {0148-0227},
 journal = {Journal of Geophysical Research},
-keywords = {Central and South America,climate change,extreme indices},
-mendeley-groups = {gosset},
 number = {D23},
 pages = {D23107},
 title = {{Changes in precipitation and temperature extremes in Central America and northern South America, 1961–2003}},
@@ -64,22 +61,20 @@ year = {2005}
 @article{DeSousa2018,
 author = {{de Sousa}, Kau{\^{e}} and Casanoves, Fernando and Sellare, Jorge and Ospina, Alejandra and Suchini, Jose Gabriel and Aguilar, Amilcar and Mercado, Leida},
 doi = {10.1016/j.jrurstud.2018.09.018},
-file = {:C\:/Users/kaued/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/de Sousa et al. - 2018 - How climate awareness influences farmers' adaptation decisions in Central America.pdf:pdf},
 issn = {07430167},
 journal = {Journal of Rural Studies},
-mendeley-groups = {agroforestry-options},
 month = {nov},
 pages = {11--19},
 publisher = {Elsevier},
 title = {{How climate awareness influences farmers' adaptation decisions in Central America?}},
-url = {https://doi.org/10.1016/j.jrurstud.2018.09.018 https://linkinghub.elsevier.com/retrieve/pii/S0743016718303425},
+url = {https://doi.org/10.1016/j.jrurstud.2018.09.018},
 volume = {64},
 year = {2018}
 }
 
 
 @article{jsonlite,
-    title = {The jsonlite Package: A Practical and Consistent Mapping Between JSON Data and R Objects},
+    title = {The jsonlite {P}ackage: {A} {P}ractical and {C}onsistent {M}apping {B}etween {JSON} {D}ata and {R} {O}bjects},
     author = {Jeroen Ooms},
     journal = {arXiv:1403.2805 [stat.CO]},
     year = {2014},
@@ -98,7 +93,7 @@ year = {2018}
 
 @article{sf,
     author = {Edzer Pebesma},
-    title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
+    title = {{{S}imple {F}eatures for {R}: {S}tandardized {S}upport for {S}patial {V}ector {D}ata}},
     year = {2018},
     journal = {{The R Journal}},
     doi = {10.32614/RJ-2018-009},
@@ -110,7 +105,7 @@ year = {2018}
   
 
 @Manual{crul,
-    title = {{crul}: HTTP Client},
+    title = {{crul}: {HTTP} {C}lient},
     author = {Scott Chamberlain},
     year = {2019},
     note = {R package version 0.7.4},
@@ -207,7 +202,7 @@ keywords = {Climate change,Diarrheal disease,Mozambique,Precipitation,Temperatur
 month = {apr},
 number = {4},
 publisher = {MDPI AG},
[…]
