+++ inst/paper/paper.bib
@@ -74,21 +74,12 @@ year = {2018}
 
 
 @article{jsonlite,
-    title = {The jsonlite {P}ackage: {A} {P}ractical and {C}onsistent {M}apping {B}etween {JSON} {D}ata and {R} {O}bjects},
+    title = {The {jsonlite} {P}ackage: {A} {P}ractical and {C}onsistent {M}apping {B}etween {JSON} {D}ata and {R} {O}bjects},
     author = {Jeroen Ooms},
     journal = {arXiv:1403.2805 [stat.CO]},
     year = {2014},
     url = {https://arxiv.org/abs/1403.2805},
 }
-
-
-@Manual{tibble,
-    title = {tibble: Simple Data Frames},
-    author = {Kirill Müller and Hadley Wickham},
-    year = {2019},
-    note = {R package version 2.1.3},
-    url = {https://CRAN.R-project.org/package=tibble},
-} 
 
 
 @article{sf,
@@ -188,7 +179,7 @@ issn = {00029637},
 journal = {American Journal of Tropical Medicine and Hygiene},
 pages = {32--45},
 publisher = {American Society of Tropical Medicine and Hygiene},
-title = {{Using rainfall and temperature data in the evaluation of national malaria control programs in Africa}},
+title = {{Using rainfall and temperature data in the evaluation of national malaria control programs in {A}frica}},
 volume = {97},
 year = {2017}
 }
@@ -210,7 +201,7 @@ year = {2018}
 @misc{Huffman2014,
 author = {Huffman, G. and Bolvin, D. and Braithwaite, D. and Hsu, K. and Joyce, R. and Xie, P.},
 publisher = {NASA's Precipitation Processing Center},
-title = {{Integrated Multi-satellitE Retrievals for GPM (IMERG), version 4.4.}},
+title = {{{I}ntegrated {M}ulti-satellit{E} {R}etrievals for {GPM} ({IMERG}), version 4.4.}},
 url = {ftp://arthurhou.pps.eosdis.nasa.gov/gpmdata/},
 year = {2014}
 }
